refactor: split CLI entrypoints and share client bootstrap
